fix: move tests classes to a dedicate package to avoid split package warning
