+++ itests/camel-k-itests-loader-jsh/src/test/java/org/apache/camel/k/loader/jsh/it/JshLoaderTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.loader.jsh;
+package org.apache.camel.k.loader.jsh.it;
 
 import java.io.IOException;
 import java.io.InputStream;
